- added __init__.py to tests so they can be located in VS-code - separated test cases for functions and classes into two files - rearranged tests using object oriented unittest.TestCase
